# api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Workflow orchestration platform for financial validation
  termsOfService: http://swagger.io/terms/
  title: Flowker API
  version: 1.0.0
servers:
- url: //localhost:4021/
paths:
...
  /v1/workflows:
    get:
      description: Retrieves a paginated list of workflows with optional filtering
      parameters:
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - draft
          - active
          - inactive
          type: string
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        schema:
          type: string
      - description: Sort field
        in: query
        name: sortBy
        schema:
          default: createdAt
          enum:
          - createdAt
          - updatedAt
          - name
          type: string
      - description: Sort order
        in: query
        name: sortOrder
        schema:
          default: DESC
          enum:
          - ASC
          - DESC
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowListOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: List workflows
      tags:
      - Workflows
    post:
      description: Creates a new workflow definition in draft status
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.CreateWorkflowInput'
        description: Workflow definition
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowCreateOutput'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Create a new workflow
      tags:
      - Workflows
      x-codegen-request-body-name: workflow
  /v1/workflows/from-template:
    post:
      description: Creates a new workflow from a registered template with provided
        parameters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.CreateWorkflowFromTemplateInput'
        description: Template ID and parameters
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowCreateOutput'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Create workflow from template
      tags:
      - Workflows
      x-codegen-request-body-name: input
  /v1/workflows/{id}:
    delete:
      description: Deletes a workflow definition (only draft and inactive workflows
        can be deleted)
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Delete a workflow
      tags:
      - Workflows
    get:
      description: Retrieves a workflow definition by its ID
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Get workflow by ID
      tags:
      - Workflows
    put:
      description: Updates an existing workflow definition (only draft workflows can
        be updated)
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.UpdateWorkflowInput'
        description: Updated workflow definition
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Update a workflow
      tags:
      - Workflows
      x-codegen-request-body-name: workflow
  /v1/workflows/{id}/activate:
    post:
      description: Transitions a workflow from draft to active status
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Activate a workflow
      tags:
      - Workflows
  /v1/workflows/{id}/clone:
    post:
      description: Creates a copy of an existing workflow with a new name
      parameters:
      - description: Source workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.CloneWorkflowInput'
        description: Clone parameters
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowCreateOutput'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Clone a workflow
      tags:
      - Workflows
      x-codegen-request-body-name: clone
  /v1/workflows/{id}/deactivate:
    post:
      description: Transitions a workflow from active to inactive status
      parameters:
      - description: Workflow ID
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowOutput'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Conflict
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Internal Server Error
      summary: Deactivate a workflow
      tags:
      - Workflows
...
components:
  schemas:
    api.ErrorResponse:
      properties:
        code:
          example: FLK-0001
          type: string
        message:
          example: Invalid input provided
          type: string
        title:
          example: Bad Request
          type: string
      required:
      - code
      - message
      - title
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.CreateWorkflowFromTemplateInput:
      properties:
        params:
          additionalProperties:
            type: object
          type: object
        templateId:
          type: string
      required:
      - params
      - templateId
      type: object
...
    github_com_LerianStudio_flowker_pkg_model.WorkflowListOutput:
      example:
        nextCursor: nextCursor
        hasMore: true
        items:
        - createdAt: createdAt
          metadata:
            key: '{}'
          nodes:
          - data:
              key: '{}'
            name: name
            id: id
            position:
              x: 0
              y: 6
            type: type
          - data:
              key: '{}'
            name: name
            id: id
            position:
              x: 0
              y: 6
            type: type
          edges:
          - sourceHandle: sourceHandle
            condition: condition
            id: id
            label: label
            source: source
            target: target
          - sourceHandle: sourceHandle
            condition: condition
            id: id
            label: label
            source: source
            target: target
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
          updatedAt: updatedAt
        - createdAt: createdAt
          metadata:
            key: '{}'
          nodes:
          - data:
              key: '{}'
            name: name
            id: id
            position:
              x: 0
              y: 6
            type: type
          - data:
              key: '{}'
            name: name
            id: id
            position:
              x: 0
              y: 6
            type: type
          edges:
          - sourceHandle: sourceHandle
            condition: condition
            id: id
            label: label
            source: source
            target: target
          - sourceHandle: sourceHandle
            condition: condition
            id: id
            label: label
            source: source
            target: target
          name: name
          description: description
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: status
          updatedAt: updatedAt
      properties:
        hasMore:
          type: boolean
        items:
          items:
            $ref: '#/components/schemas/github_com_LerianStudio_flowker_pkg_model.WorkflowOutput'
          type: array
        nextCursor:
          type: string
      type: object
...